pyproject.toml
[project]
name = "uicdev-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.6.3",
    "django>=5.2.12",
    "django-celery>=3.1.17",
    "django-celery-beat>=2.9.0",
    "django-jazzmin>=3.0.4",
    "django-modeltranslation>=0.20.2",
    "django-redis>=6.0.0",
    "django-rosetta>=0.10.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "fastapi>=0.136.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
]
